Well it works for Intel Macs right now
CG.Framework is a hard dependency from NVidia.  We can replace it with GLSL but from all projects that did this, they have not worked.  But it is playable on Rosetta 2 for M1 users.

+++ HPL/sources/math/CRC.h
@@ -19,12 +19,10 @@
 #ifndef HPL_CRC_H
 #define HPL_CRC_H
 
-#include <stdint.h>
-
 namespace hpl {
 	//----------------------------------------
 
-	typedef uintptr_t tCRCKey;
+	typedef unsigned long tCRCKey;
 
 	//----------------------------------------
